fix #719, adding random_generator in dedicated file
