+++ pdi/include/pdi/testing.h
@@ -45,7 +45,7 @@
  * Each test will run in a separate directory that's cleaned up after execution, feel free to create
  * any file you need there.
  * In order to be able to use make_a for your own classes & structures, you'll need to define a
- * matching random_init() function.
+ * matching random_init() function (These function are defined in random_generator.h).
  * 
  * Some recommendations:
  * - Follow gtest naming requirements for the tests: valid C++ identifiers without _ (underscores)
@@ -64,110 +64,16 @@
 
 #include <algorithm>
 #include <filesystem>
-#include <random>
-#include <ranges>
 #include <type_traits>
 
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
 
 #include <paraconf.h>
+#include <pdi/random_generator.h>
 #include <pdi.h>
 
 namespace PDI {
-
-/** An object is initializable_from<G> if it can be initalized using the init_from member function
- * from a reference of G
- */
-template < class T, class G >
-concept initializable_from = requires(T t, G& gen) {
-	{
-		t.init_from(gen)
-	};
-};
-
-/** An object is buildable_from<G> if it can be constructed using the init_from member function
- * from a reference of G
- */
-template <class T, class G>
-concept buildable_from = std::constructible_from<T, G&> && requires(G& g) {
-	{
-		T(g)
-	};
-	{
-		::new T(g)
-	};
-};
-
-/** initialize an integral value with uniformly distributed values from the provided generator
- * \param gen the (pseudo)random generator
- * \param t the value to initialize
- */
-static inline void random_init(std::uniform_random_bit_generator auto& gen, std::integral auto& t)
-{
-	using T = std::remove_reference_t<decltype(t)>;
-	t = std::uniform_int_distribution<T>(std::numeric_limits<T>::min())(gen);
-}
-
-/** initialize a floating-point value with normally distributed values from the provided generator
- * \param gen the (pseudo)random generator
- * \param t the value to initialize
- */
-static inline void random_init(std::uniform_random_bit_generator auto& gen, std::floating_point auto& t)
-{
-	t = std::normal_distribution<std::remove_reference_t<decltype(t)>>{}(gen);
-}
-
-/** initialize all elements in an iterable with the provided generator
- * \param gen the (pseudo)random generator
- * \param t the iterable to initialize
- */
-static inline void random_init(std::uniform_random_bit_generator auto& gen, std::ranges::input_range auto& t)
-{
-	std::for_each(std::begin(t), std::end(t), [&](std::ranges::range_value_t<decltype(t)>& v) { random_init(gen, v); });
-}
-
-/** initialize an object that supports `init_from` from the provided generator
- * \param gen the (pseudo)random generator
- * \param t the object to initialize
- */
-static inline void random_init(std::uniform_random_bit_generator auto& gen, initializable_from<decltype(gen)> auto& t)
-{
-	t.init_from(gen);
-}
-
-/** initialize an object that can be constructed from a generator from the provided generator
- * \param gen the (pseudo)random generator
- * \param t the object to initialize
- */
-template <typename T, std::uniform_random_bit_generator G>
-requires(!std::ranges::input_range<T> && !initializable_from<T, G> && buildable_from<T, G> && std::assignable_from<T&, T &&>)
-static inline void random_init(G& gen, T& t)
-{
-	t = T(gen);
-}
-
-/** make a new randomly-initialized object that can be constructed from a generator
- * \param gen the (pseudo)random generator
- */
-template <typename T, std::uniform_random_bit_generator G>
-requires buildable_from<T, G>
-static inline T make_random(G& gen)
-{
-	return T(gen);
-}
-
-/** make a new randomly-initialized object that supports `random_init`
- * \param gen the (pseudo)random generator
- */
-template <typename T, std::uniform_random_bit_generator G>
-requires(!buildable_from<T, G>, std::default_initializable<T>)
-static inline T make_random(G& gen)
-{
-	T result;
-	random_init(gen, result);
-	return result;
-}
 
 /** The fixture class for PDI plugins testing.
  *
